--- Assets/Follow.cs
@@ -1,53 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Follow : MonoBehaviour
-{
-    //Public variable to store a reference to the player game object
-    public GameObject player;
-
-
-    //Private variable to store the offset distance between the player and camera
-    private Vector3 offset;
-
-    public Material trans;
-
-    public Material originalMat;
-
-    public Transform camObject;
-
-    void UseTransparentMaterial(Transform thing)
-    {
-        Renderer thingRenderer;
-        thingRenderer = thing.GetComponent<Renderer>();
-        thingRenderer.sharedMaterial = trans;
-    }
-
-    void UseNormalMaterial(Transform thing)
-    {
-        Renderer thingRenderer;
-        thingRenderer = thing.GetComponent<Renderer>();
-        thingRenderer.sharedMaterial = originalMat;
-    }
-
-    void CameraClipingDetection()
-    {
-        RaycastHit hit;
-        Debug.DrawLine(player.transform.position, transform.position, Color.red, 0.1f);
-    }
-
-    // Use this for initialization
-    void Start()
-    {
-        //Calculate and store the offset value by getting the distance between the player's position and camera's position.
-        offset = transform.position - player.transform.position;
-    }
-
-    // LateUpdate is called after Update each frame
-    void LateUpdate()
-    {
-        // Set the position of the camera's transform to be the same as the player's, but offset by the calculated offset distance.
-        transform.position = player.transform.position + offset;
-    }
-}
