--- Assets/Prefabs/Company Building.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: a942caeb5f740f04699fa9264dd1cc43
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
