+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2019.2.6f1
-m_EditorVersionWithRevision: 2019.2.6f1 (fe82a0e88406)
+m_EditorVersion: 2019.2.9f1
+m_EditorVersionWithRevision: 2019.2.9f1 (ebce4d76e6e8)
